move tests to credential specific modules
