+++ sdk/identity/azure-identity/tests/test_vscode_credential.py
@@ -93,7 +93,7 @@ def test_no_obtain_token_if_cached():
 
     mock_client = mock.Mock(spec=object)
     mock_client.obtain_token_by_refresh_token = mock.Mock(return_value=expected_token)
-    mock_client.get_cached_access_token = mock.Mock(return_value='VALUE')
+    mock_client.get_cached_access_token = mock.Mock(return_value="VALUE")
 
     with mock.patch(VSCodeCredential.__module__ + ".get_credentials", return_value="VALUE"):
         credential = VSCodeCredential(_client=mock_client)
